[Fix] 보상 받기 시 상태 수정 정상화
